requirements.txt
# Django
Django==4.2
djangorestframework==3.15.1
django-filter==24.2

# Photo & Image
pillow==10.3.0

# Authentication
djangorestframework-simplejwt==5.3.1